WIP: works now.

+++ deep_sort/sort/detection.py
@@ -27,7 +27,9 @@ class Detection(object):
     """
 
     def __init__(self, tlwh, confidence, feature):
-        self.tlwh = np.asarray(tlwh, dtype=np.float)
+        # print(tlwh)
+        self.tlwh = np.asarray(tlwh, dtype=np.float32)
+        # self.tlwh = np.asarray(tlwh)
         self.confidence = float(confidence)
         self.feature = np.asarray(feature, dtype=np.float32)
